processing_python/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
pydantic
pillow
pytesseract
python-multipart
qdrant-client
langchain
langchain-google-genai
langchain-core
torch
torchvision
numpy
pyyaml
tqdm
opencv-python-headless
ml-collections